dts: nordic: add nRF53 GPIO compatible and security property
The GPIO hardware block on the nRF53 has split security. In non-secure
mode, it can be accessed by both secure and non-secure accesses. It
will be useful to allow the user to configure a bitmask of pins to
configure as secure, to prevent non-secure accesses to some pins.

Since this is a hardware difference from nRF52 GPIO controllers, add a
new GPIO compatible, "nordic,nrf53-gpio", for this SoC's GPIO block.
In the binding for that new compatible, add a "secure-pin-mask" that
can be used to provide a bit mask of GPIOs which should be configured
as secure at boot time.

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>